requirements.txt
# RAG assignment dependencies. Add/version as needed.
# Note: You can use any other libraries that are not listed here
sentence-transformers
openai
flask
flask-cors
pyyaml
numpy
google-generativeai
# or: fastapi uvicorn
# optional: pyyaml for config
